+++ actix-http/CHANGES.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Parse all HTTP/1 `Connection` header options consistently. [#2692]
-- Only treat `Expect: 100-continue` as a continue expectation.
+- Only treat `Expect: 100-continue` as a continue expectation in HTTP/1.1 requests.
 
 [#2692]: https://github.com/actix/actix-web/issues/2692